move Header implementation out to its own file
